requirements.txt
# requirements.txt
#torch>=2.0.0
#flash-attn 
transformers==4.49.0
datasets==3.6.0
huggingface_hub==0.30.2
gradio==5.29.0
pillow==11.2.1
tqdm==4.67.1
matplotlib==3.10.1
numpy==1.26.4
timm==1.0.15
einops==0.8.1
opencv-python==4.11.0.86